Refactor status form into reusable components
